+++ calva/utilities.ts
@@ -7,10 +7,10 @@ import * as path from 'path';
 import { NReplSession } from './nrepl';
 import { activeReplWindow } from './repl-window';
 const syntaxQuoteSymbol = "`";
-const { parseForms } = require('../cljs-out/cljs-lib');
+const { parseForms } = require('../out/cljs-lib/cljs-lib');
 import * as docMirror from './calva-fmt/ts/docmirror';
-import { TokenCursor, LispTokenCursor } from './webview-src/client/token-cursor';
-import { Token } from './webview-src/client/clojure-lexer';
+import { TokenCursor, LispTokenCursor } from './webview/token-cursor';
+import { Token } from './webview/clojure-lexer';
 import select from './select';